Disable pipefail
With pipefail the exit status of a pipe will be non-zero if any
component exits non-zero. All well-behaved components will exit non-zero
when their output handle is closed before all the component's content
could be written.

This problem affects both pipes in the program. It seems to be less
problematic to just disable this sanity check here.