Add two helper scripts
The script build-kernel.sh encodes the current knowledge about how to
compile a Mainline kernel. Especially tricky is the part of generating a
valid config.

The second script is incredible helpful when tracing down which commit
breaks the rtl88x2bu build. Knowing the cause very often helped to
fix the driver.